Fix Socket::VERSION evaluation
The Net::Ping module evaluates the $Socket::VERSION variable and then
assigns the result to $Socket_VERSION. This is done so if the version
contains an underscore (e.g. '2.020_03') it gets properly converted to a
number. $Socket_VERSION can be then compared without errors/warnings
with numeric literal versions in the rest of the code. The conversion is
however currently attempted using block eval which results in the same
version string without converting it to a number. The patch fixes the
problem by using string eval.